Extract VariableReplacerFixture base class
